# public/openapi.yaml
openapi: 3.1.0
info:
  title: Starfagrunnur API
  description: |
    Open Icelandic occupational knowledge graph. Combines ÍSTARF21 (the Icelandic
    ISCO-08 adaptation) categorisation with ESCO occupations and skills.

    **Data model** (see the graph for the full picture):
    ```
    (OccupationGroup) -[:CHILD_OF]-> (OccupationGroup)     # 4-level ÍSTARF21 hierarchy
    (EscoOccupation) -[:BELONGS_TO_GROUP]-> (OccupationGroup)
    (EscoOccupation) -[:REQUIRES_SKILL {relationType}]-> (Skill)
    (AliasTerm) -[:ALIAS_OF]-> (EscoOccupation)
    ```

    **ÍSTARF21 levels**:
    - Level 1 – bálkur (major group, 1-digit) — 9 categories
    - Level 2 – deild (sub-major, 2-digit) — 39 categories
    - Level 3 – klasi (minor, 3-digit) — 121 categories
    - Level 4 – starfaflokkur (unit, 4-digit) — 409 categories; this is where ESCO roles attach and tasks live

    **Skill classification (ESCO semantics)**:
    - `relationType: essential` — typical for the occupation across the European labour market
    - `relationType: optional` — may be required depending on specialisation/context
    - `skillType: skill/competence` — doing skills
    - `skillType: knowledge` — knowing things

    All labels are in Icelandic (is-IS).
  version: "1.0.0"
  contact:
    name: Peritus
    url: https://starfagrunnur.is
servers:
  - url: /
    description: Current host (the origin serving this documentation)
  - url: https://starfagrunnur.is
    description: Production
...
    description: API alias
...
paths:
...
  /api/v1/occupations:
    get:
      tags: [Occupations]
      summary: List ÍSTARF21 occupation groups
      description: |
        Browse/list groups. Filter by level and/or parent code. Paginated.

        **Answers questions like:**
        - "What are the top-level occupation categories in Iceland?" → `?level=1`
        - "List every starfaflokkur (level-4 unit)." → `?level=4`
        - "Which subcategories live directly under bálkur 2?" → `?parent=2`
        - "I want to build a category browser / dropdown."
      parameters:
        - in: query
          name: level
          schema: { type: integer, enum: [1, 2, 3, 4] }
          description: Filter by hierarchy level.
        - in: query
          name: parent
          schema: { type: string, pattern: "^\\d{1,4}$" }
          description: Return only direct children of this parent code.
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 1000 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/OccupationSummary" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/occupations/{code}:
    get:
      tags: [Occupations]
      summary: Get an ÍSTARF21 occupation group
      description: |
        Full detail for a single group: title, description, tasks (level-4 only), parent.
        Tasks appear at level-4 unit groups only (by ÍSTARF21 design). Levels 1–3 are
        aggregated categories and their `tasks` will be `null`.

        **Answers questions like:**
        - "What is occupation 2221 about?"
        - "Give me the official tasks and description for this ÍSTARF code."
        - "What category does 2221 sit directly under?" (via `parent`)
      parameters:
        - $ref: "#/components/parameters/Code"
      responses:
        "200":
          description: Occupation group detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccupationDetail" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/occupations/{code}/descendants:
    get:
      tags: [Occupations]
      summary: All descendants
      description: |
        All descendants of this group, optionally filtered to a specific level.

        **Answers questions like:**
        - "List every starfaflokkur (level 4) under bálkur 2." → `?level=4`
        - "Give me the whole subtree below this klasi."
        - "How many level-4 units live under *Sérfræðistörf*?"
      parameters:
        - $ref: "#/components/parameters/Code"
        - in: query
          name: level
          schema: { type: integer, enum: [2, 3, 4] }
          description: Return only descendants at this level.
      responses:
        "200":
          description: Descendants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OccupationSummary" }
...
  /api/v1/occupations/{code}/esco:
    get:
      tags: [Occupations]
      summary: ESCO occupations under this group
      description: |
        ESCO roles belonging to this ÍSTARF21 group.

        - Without `recursive`: only roles attached directly (non-empty on level-4 groups).
        - With `recursive=true`: walks down to every level-4 descendant and unions their
          ESCO roles. Use this to get "all roles under *Sérfræðistörf*" from bálkur code `2`.

        **Answers questions like:**
        - "Which concrete job titles live under ÍSTARF category 2221?" (level 4)
        - "Give me every ESCO role under the healthcare bálkur 2." → `?recursive=true`
        - "What are the actual job names a person can have inside this category?"
        - ÍSTARF categorises; ESCO names the roles — this endpoint bridges the two.
      parameters:
        - $ref: "#/components/parameters/Code"
        - in: query
          name: recursive
          schema: { type: boolean, default: false }
          description: Include ESCO roles from descendant level-4 groups.
      responses:
        "200":
          description: ESCO occupations in/under the group
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/EscoOccupationSummary"
                    - type: object
                      properties:
                        group_code:
                          type: string
                          description: Only present when recursive=true — the level-4 group the role attaches to.
                        group_title:
                          type: string
                          description: Only present when recursive=true.
...
  /api/v1/esco-occupations:
    get:
      tags: [ESCO]
      summary: List ESCO occupations
      description: |
        Enumerate / search ESCO occupations. Without `q`, returns all occupations
        sorted alphabetically (filterable, paginated). With `q`, returns the
        top-ranked full-text matches.

        **Answers questions like:**
        - "List every ESCO role in ISCO-08 group 2221." → `?isco_group=2221`
        - "Find ESCO roles matching *hjúkr*." → `?q=hjúkr`
        - "Give me all roles under the healthcare bálkur." → `?in_istarf21=2`
        - "Enumerate all 3,039 ESCO occupations (paginated)." → no params, paginate with `offset`
      parameters:
        - in: query
          name: q
          schema: { type: string, minLength: 2 }
          description: Full-text search over Icelandic labels.
        - in: query
          name: isco_group
          schema: { type: string, pattern: "^\\d{1,4}$" }
          description: Filter by ISCO-08 group (e.g. 2221).
        - in: query
          name: in_istarf21
          schema: { type: string, pattern: "^\\d{1,4}$" }
          description: Filter to ESCO roles under this ÍSTARF21 code or any of its descendants.
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of ESCO occupations
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        uuid: { type: string }
                        conceptUri: { type: string }
                        preferredLabel_is: { type: string }
                        description_is: { type: [string, "null"] }
                        iscoGroup: { type: string }
                        group_code: { type: [string, "null"] }
                        group_title: { type: [string, "null"] }
  /api/v1/esco-occupations/{uuid}:
    get:
      tags: [ESCO]
      summary: Get an ESCO occupation
      description: |
        Full detail for a single ESCO occupation: description, ISCO-08 group,
        alias labels, parent ÍSTARF21 group with full ancestor chain, and
        counts of essential/optional skills.

        **Answers questions like:**
        - "Tell me everything about the *hjúkrunarfræðingur* role."
        - "What category does this ESCO role fit into in the Icelandic classification?"
        - "How many skills does this role have (without fetching them all)?"
        - "What alternative labels / aliases exist for this occupation?"
      parameters:
        - $ref: "#/components/parameters/EscoUuid"
      responses:
        "200":
          description: ESCO occupation detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EscoOccupationDetail" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/esco-occupations/{uuid}/skills:
    get:
      tags: [ESCO]
      summary: Skills for this ESCO occupation
      description: |
        Paginated list of skills required by one specific ESCO occupation.

        **Answers questions like:**
        - "What skills does a *hjúkrunarfræðingur* actually need?"
        - "Show me only the essential skills for this role." → `?relation_type=essential`
        - "I want all ~90 skills for this role, paginated." → iterate with `page`
        - Precision-oriented alternative to `/occupations/{code}/skills`, which unions across the whole group.
      parameters:
        - $ref: "#/components/parameters/EscoUuid"
        - in: query
          name: relation_type
          schema: { type: string, enum: [essential, optional] }
        - in: query
          name: skill_type
          schema: { type: string, enum: ["skill/competence", "knowledge"] }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Paginated skills
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillsResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/skills/{uuid}/occupations:
    get:
      tags: [Skills]
      summary: Occupations requiring this skill (reverse lookup)
      description: |
        All ESCO occupations that have a `REQUIRES_SKILL` edge to this skill,
        with the relation type and parent ÍSTARF21 group. Paginated.

        **Answers questions like:**
        - "Which jobs require *Python*?"
        - "List only roles where this skill is *essential*, not optional." → `?relation_type=essential`
        - "What career paths open up if I learn this skill?"
        - "If I'm reskilling an employee with X skill, which roles could they fit?"
        - The core reverse-lookup endpoint for reskilling, CV-matching, and learning-path applications.
      parameters:
        - $ref: "#/components/parameters/SkillUuid"
        - in: query
          name: relation_type
          schema: { type: string, enum: [essential, optional] }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Paginated list of ESCO occupations requiring this skill
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        uuid: { type: string }
                        conceptUri: { type: string }
                        preferredLabel_is: { type: string }
                        iscoGroup: { type: string }
                        relationType: { type: string, enum: [essential, optional] }
                        group_code: { type: [string, "null"] }
                        group_title: { type: [string, "null"] }
                  total: { type: integer }
                  page: { type: integer }
                  page_size: { type: integer }
                  total_pages: { type: integer }
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  parameters:
    Code:
      name: code
      in: path
      required: true
      schema: { type: string, pattern: "^\\d{1,4}$" }
      description: ÍSTARF21 code (1–4 digits). Length determines level.
      examples:
        balkur:   { value: "2",    summary: "Level 1 — Sérfræðistörf" }
        deild:    { value: "22",   summary: "Level 2 — Sérfræðistörf í heilbrigðisgreinum" }
        klasi:    { value: "222",  summary: "Level 3" }
        unit:     { value: "2221", summary: "Level 4 — Sérfræðistörf við hjúkrun" }
    EscoUuid:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: |
        UUID part of the ESCO conceptUri. The full URI has the form
        `http://data.europa.eu/esco/occupation/<uuid>`; pass just the UUID.
      example: "8d3e8aaa-791b-4c75-a465-f3f827028f50"
    SkillUuid:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: |
        UUID part of the ESCO skill conceptUri. The full URI has the form
        `http://data.europa.eu/esco/skill/<uuid>`; pass just the UUID.
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
  schemas:
    OccupationSummary:
      type: object
      required: [code, title, level]
      properties:
        code:   { type: string, example: "2221" }
        title:  { type: string, example: "Sérfræðistörf við hjúkrun" }
        level:  { type: integer, minimum: 1, maximum: 4 }
    OccupationDetail:
      type: object
      required: [code, title, level, source_system]
      properties:
        code:          { type: string }
        title:         { type: string }
        level:         { type: integer, minimum: 1, maximum: 4 }
        source_system: { type: string, example: "ISTARF21" }
        parent:
          oneOf:
            - $ref: "#/components/schemas/OccupationSummary"
            - type: "null"
        description:    { type: [string, "null"] }
        tasks:
          description: List of tasks (ÍSTARF21 descriptions). Only populated on level-4 units; null otherwise.
          oneOf:
            - type: array
              items: { type: string }
            - type: "null"
        example_titles:
          oneOf:
            - type: array
              items: { type: string }
            - type: "null"
        notes:          { type: [string, "null"] }
    EscoOccupationSummary:
      type: object
      required: [conceptUri, preferredLabel_is, iscoGroup]
      properties:
        conceptUri:        { type: string, format: uri }
        preferredLabel_is: { type: string, example: "hjúkrunarfræðingur" }
        description_is:    { type: [string, "null"] }
        iscoGroup:         { type: string, example: "2221" }
    EscoOccupationDetail:
      type: object
      required: [conceptUri, uuid, preferredLabel_is, iscoGroup]
      properties:
        conceptUri:        { type: string, format: uri }
        uuid:              { type: string }
        preferredLabel_is: { type: string }
        description_is:    { type: [string, "null"] }
        iscoGroup:         { type: string }
        aliases:
          type: array
          items: { type: string }
          description: Alternate Icelandic labels for this occupation.
        skills:
          type: object
          properties:
            essential_count: { type: integer }
            optional_count:  { type: integer }
        group:
          description: Parent ÍSTARF21 group + ancestor chain. Null if the ESCO role isn't bridged.
          oneOf:
            - type: "null"
            - type: object
              required: [code, title, level]
              properties:
                code:      { type: string }
                title:     { type: string }
                level:     { type: integer }
                ancestors:
                  type: array
                  items: { $ref: "#/components/schemas/OccupationSummary" }
...
    SkillItem:
      type: object
      required: [preferredLabel_is, skillType, relationType]
      properties:
        preferredLabel_is: { type: string }
        skillType:
          type: string
          enum: ["skill/competence", "knowledge"]
        description_is: { type: [string, "null"] }
        relationType:
          type: string
          enum: [essential, optional]
        occupation:
          type: string
          description: Present only on the group-level /skills endpoint — names the ESCO role requiring the skill.
    SkillsResponse:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SkillItem" }
        total:       { type: integer }
        page:        { type: integer }
        page_size:   { type: integer }
        total_pages: { type: integer }
...